+++ installation-bundle/CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ### 1.2.0-beta1 (2016-XX-XX)
 
+ * Show the console output if a post installation task cannot be completed (see #32).
  * Make the legacy configuration files optional (see contao/core-bundle#521).